refactor: 상수 처리
