+++ src/Pages/Main/ConstantMain.ts
@@ -1,0 +1,15 @@
+const TEXT = {
+  EXPLAIN: `별점과 리뷰를 바탕으로 선정한\n 믿고 보는 맛집 리스트\n\n포도플레이트가 꼽은\n 특별한 맛집을 만나보세요.`,
+  GANGNAM: "믿고 보는 강남 맛집",
+  HONGDAE: "홍대입구 맛집 스토리",
+  YEOKGOK: "에디터가 선정한 역곡 식당",
+  ERROR: "오류가 발생했습니다",
+  BTN: "전체보기",
+};
+
+const KEYWORDS = {
+  LIST_KEYWORD: "강남",
+  STORY_KEYWORD: "홍대입구",
+};
+
+export default { TEXT, KEYWORDS };
